Move abstract Command.java for reuse
