+++ src/main/java/bots/manker/commands/FourTwentyCommand.java
@@ -1,5 +1,7 @@
 package bots.manker.commands;
 
+import bots.Command;
+
 public class FourTwentyCommand extends Command {
     public static final String COMMAND_NAME = "420";
