+++ src/main/java/bots/manker/commands/MemeCommand.java
@@ -1,5 +1,7 @@
 package bots.manker.commands;
 
+import bots.Command;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
